+++ compiler/rustc_interface/src/callbacks.rs
@@ -12,11 +12,10 @@
 use std::fmt;
 
 use rustc_errors::{DiagInner, TRACK_DIAGNOSTIC};
-use rustc_middle::dep_graph::{DepNodeExt, TaskDepsRef};
+use rustc_middle::dep_graph::dep_node::default_dep_kind_debug;
+use rustc_middle::dep_graph::{DepContext, DepKind, DepNode, DepNodeExt, TaskDepsRef};
 use rustc_middle::ty::tls;
 use rustc_query_impl::QueryCtxt;
-use rustc_query_system::dep_graph::dep_node::default_dep_kind_debug;
-use rustc_query_system::dep_graph::{DepContext, DepKind, DepNode};
 
 fn track_span_parent(def_id: rustc_span::def_id::LocalDefId) {
     tls::with_context_opt(|icx| {
@@ -107,9 +106,9 @@ pub fn dep_node_debug(node: DepNode, f: &mut std::fmt::Formatter<'_>) -> std::fm
 pub fn setup_callbacks() {
     rustc_span::SPAN_TRACK.swap(&(track_span_parent as fn(_)));
     rustc_hir::def_id::DEF_ID_DEBUG.swap(&(def_id_debug as fn(_, &mut fmt::Formatter<'_>) -> _));
-    rustc_query_system::dep_graph::dep_node::DEP_KIND_DEBUG
+    rustc_middle::dep_graph::dep_node::DEP_KIND_DEBUG
         .swap(&(dep_kind_debug as fn(_, &mut fmt::Formatter<'_>) -> _));
-    rustc_query_system::dep_graph::dep_node::DEP_NODE_DEBUG
+    rustc_middle::dep_graph::dep_node::DEP_NODE_DEBUG
         .swap(&(dep_node_debug as fn(_, &mut fmt::Formatter<'_>) -> _));
     TRACK_DIAGNOSTIC.swap(&(track_diagnostic as _));
 }
