+++ compiler/rustc_middle/src/dep_graph/query.rs
