+++ compiler/rustc_middle/src/query/mod.rs
@@ -1,4 +1,6 @@
+use rustc_data_structures::jobserver::Proxy;
 use rustc_hir::def_id::LocalDefId;
+use rustc_query_system::query::QuerySideEffect;
 
 pub use self::caches::{
     DefIdCache, DefaultCache, QueryCache, QueryCacheKey, SingleCache, VecCache,
@@ -10,6 +12,7 @@ pub use self::plumbing::{
     TyCtxtAt, TyCtxtEnsureDone, TyCtxtEnsureOk,
 };
 pub use self::stack::{QueryStackDeferred, QueryStackFrame, QueryStackFrameExtra};
+use crate::dep_graph::{DepNodeIndex, HasDepContext, SerializedDepNodeIndex};
 pub use crate::queries::Providers;
 use crate::ty::TyCtxt;
 
@@ -32,4 +35,19 @@ pub fn describe_as_module(def_id: impl Into<LocalDefId>, tcx: TyCtxt<'_>) -> Str
     } else {
         format!("module `{}`", tcx.def_path_str(def_id))
     }
+}
+
+pub trait QueryContext<'tcx>: HasDepContext {
+    /// Gets a jobserver reference which is used to release then acquire
+    /// a token while waiting on a query.
+    fn jobserver_proxy(&self) -> &Proxy;
+
+    /// Load a side effect associated to the node in the previous session.
+    fn load_side_effect(
+        self,
+        prev_dep_node_index: SerializedDepNodeIndex,
+    ) -> Option<QuerySideEffect>;
+
+    /// Register a side effect for the given node, for use in next session.
+    fn store_side_effect(self, dep_node_index: DepNodeIndex, side_effect: QuerySideEffect);
 }
